refactored all(moved Action namespace to Rule/Action)
