+++ src/SAREhub/EasyECA/Rule/Action/ActionDefinition.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SAREhub\EasyECA\Action;
+namespace SAREhub\EasyECA\Rule\Action;
 
 use OutOfBoundsException;
