+++ tests/unit/SAREhub/EasyECA/Rule/Action/ActionParserTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace SAREhub\EasyECA\Action;
+namespace SAREhub\EasyECA\Rule\Action;
 
 
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
